refactor(cards): extract ExplorerPortalMenu, highlight, banner, and footer subcomponents
