+++ src/components/ProfileCard.tsx
@@ -1,176 +1,21 @@
 'use client';
 
 import Image from 'next/image';
-import { NDKEvent, NDKUser } from '@nostr-dev-kit/ndk';
+import { NDKEvent } from '@nostr-dev-kit/ndk';
 import AuthorBadge from '@/components/AuthorBadge';
-import { getNewestProfileMetadata, getNewestProfileEvent } from '@/lib/vertex';
+import { getNewestProfileEvent } from '@/lib/vertex';
 import { isAbsoluteHttpUrl } from '@/lib/urlPatterns';
 import { useEffect, useMemo, useState, useRef } from 'react';
 import { useRouter, usePathname } from 'next/navigation';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faExternalLink, faArrowLeft, faBoltLightning, faHouseUser, faArrowUpRightFromSquare, faCode, faMobileScreenButton } from '@fortawesome/free-solid-svg-icons';
-import TitleBarButton from '@/components/TitleBarButton';
 import CopyButton from '@/components/CopyButton';
 import { shortenNpub, trimImageUrl, calculateAbsoluteMenuPosition } from '@/lib/utils';
 import { nip19 } from 'nostr-tools';
 import { setPrefetchedProfile, prepareProfileEventForPrefetch } from '@/lib/profile/prefetch';
 import { createProfileExplorerItems } from '@/lib/portals';
 import RawEventJson from '@/components/RawEventJson';
-import CardActions from '@/components/CardActions';
-import { formatRelativeTimeAuto, formatExactDate } from '@/lib/relativeTime';
-import Nip05Display from '@/components/Nip05Display';
-import { useHasSentZap, useHasSentNutzap } from '@/hooks/useHasSentZap';
-import { createPortal } from 'react-dom';
-
-// Import centralized URL utilities
-import { cleanWebsiteUrl } from '@/lib/utils/urlUtils';
-
-function cleanLightningAddress(lightning: string, npub: string): string {
-  // If lightning address starts with the user's npub, remove it
-  if (lightning.startsWith(npub)) {
-    return lightning.substring(npub.length);
-  }
-  return lightning;
-}
-
-function ProfileCreatedAt({ pubkey, fallbackEventId, fallbackCreatedAt, lightning, website, npub, onToggleRaw, user, onAuthorClick, onToggleMenu, menuButtonRef }: { pubkey: string; fallbackEventId?: string; fallbackCreatedAt?: number; lightning?: string; website?: string; npub: string; onToggleRaw: () => void; user: NDKUser; onAuthorClick?: (npub: string) => void; onToggleMenu?: () => void; menuButtonRef?: React.RefObject<HTMLButtonElement | null> }) {
-  const [updatedAt, setUpdatedAt] = useState<number | null>(null);
-  const [updatedEventId, setUpdatedEventId] = useState<string | null>(null);
-  const router = useRouter();
-  const pathname = usePathname();
-  const hasSentZap = useHasSentZap(pubkey);
-  const hasSentNutzap = useHasSentNutzap(pubkey);
-  const lightningButtonAccent = hasSentZap && hasSentNutzap ? 'text-green-400' : hasSentZap ? 'text-yellow-200' : hasSentNutzap ? 'text-purple-400' : '';
-  const lightningIconAccent = hasSentZap && hasSentNutzap ? 'text-green-400' : hasSentZap ? 'text-yellow-200' : hasSentNutzap ? 'text-purple-400' : '';
-  const lightningAnchorAccent = hasSentZap && hasSentNutzap
-    ? 'text-green-400 hover:text-green-300'
-    : hasSentZap
-      ? 'text-yellow-200 hover:text-yellow-100'
-      : hasSentNutzap
-        ? 'text-purple-400 hover:text-purple-300'
-        : 'text-gray-400 hover:text-gray-200';
-
-  const handleLightningSearch = (e: React.MouseEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (!lightning) return;
-    const searchQuery = `(kind:0 OR kind:1) ${lightning}`;
-    const params = new URLSearchParams();
-    params.set('q', searchQuery);
-    router.push(`/?${params.toString()}`);
-  };
-
-  const handleWebsiteSearch = (e: React.MouseEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (!website) return;
-    const params = new URLSearchParams();
-    params.set('q', website);
-    router.push(`/?${params.toString()}`);
-  };
-
-  useEffect(() => {
-    let isMounted = true;
-    (async () => {
-      try {
-        const newest = await getNewestProfileMetadata(pubkey);
-        if (!isMounted) return;
-        if (newest) { setUpdatedAt(newest.created_at || null); setUpdatedEventId(newest.id || null); }
-      } catch {
-        if (!isMounted) return;
-        setUpdatedAt(fallbackCreatedAt || null);
-        setUpdatedEventId(fallbackEventId || null);
-      }
-    })();
-    return () => { isMounted = false; };
-  }, [pubkey, fallbackEventId, fallbackCreatedAt]);
-
-  const updatedLabel = updatedAt ? formatRelativeTimeAuto(updatedAt) : 'Unknown';
-  const cleanedLightning = lightning ? cleanLightningAddress(lightning, npub) : undefined;
-  const timestampProps = typeof updatedAt === 'number'
-    ? { 'data-timestamp': String(updatedAt) }
-    : {};
-
-  return (
-    <div className="text-xs text-gray-300 bg-[#2d2d2d] border-t border-[#3d3d3d] px-4 py-2 flex items-center gap-3 flex-wrap">
-      <div className="flex items-center gap-2 min-h-[1rem]">
-        {user && <Nip05Display user={user} onProfileClick={onAuthorClick} />}
-        {lightning ? (
-          <div className="inline-flex items-center gap-1">
-            <button
-              type="button"
-              onClick={handleLightningSearch}
-              className={`inline-flex items-center gap-1 hover:underline p-1 rounded ${lightningButtonAccent}`.trim()}
-              title={`Search for ${lightning}`}
-            >
-              <FontAwesomeIcon icon={faBoltLightning} className={`h-4 w-4 ${lightningIconAccent}`.trim()} />
-              <span className="truncate max-w-[14rem] hidden sm:inline">{cleanedLightning}</span>
-            </button>
-            <a
-              href={`lightning:${lightning}`}
-              className={`p-1 rounded hover:bg-gray-600 hidden sm:block ${lightningAnchorAccent}`.trim()}
-              title={`Open ${lightning} in Lightning wallet`}
-              onClick={(e) => e.stopPropagation()}
-            >
-              <FontAwesomeIcon icon={faExternalLink} className={`h-4 w-4 ${lightningIconAccent}`.trim()} />
-            </a>
-          </div>
-        ) : null}
-        {website && isAbsoluteHttpUrl(website) ? (
-          <span className="inline-flex items-center gap-1">
-            <button
-              type="button"
-              onClick={handleWebsiteSearch}
-              className="inline-flex items-center gap-1 hover:underline p-1 rounded"
-              title={`Search for ${website}`}
-            >
-              <FontAwesomeIcon icon={faHouseUser} className="h-4 w-4" />
-              <span className="truncate max-w-[14rem] hidden sm:inline">{cleanWebsiteUrl(website)}</span>
-            </button>
-            <a
-              href={website}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-gray-400 hover:text-gray-200 p-1 rounded hover:bg-gray-600 hidden sm:block"
-              title={`Open ${website} externally`}
-              onClick={(e) => e.stopPropagation()}
-            >
-              <FontAwesomeIcon icon={faExternalLink} className="h-4 w-4" />
-            </a>
-          </span>
-        ) : null}
-        {/* NIP-05 controls moved to AuthorBadge next to the name */}
-      </div>
-      <div className="ml-auto flex items-center gap-2">
-        <div className="flex items-center gap-2">
-          {updatedAt && updatedEventId ? (
-            pathname.startsWith('/p/') ? (
-              <button
-                onClick={onToggleRaw}
-                className="hover:underline cursor-pointer"
-                title={updatedAt ? formatExactDate(updatedAt) : undefined}
-                {...timestampProps}
-              >
-                {updatedLabel}
-              </button>
-            ) : (
-              <a href={`/p/${npub}`} className="hover:underline" title={updatedAt ? formatExactDate(updatedAt) : undefined} {...timestampProps}>{updatedLabel}</a>
-            )
-          ) : (
-            <span title={updatedAt ? formatExactDate(updatedAt) : undefined} {...timestampProps}>{updatedLabel}</span>
-          )}
-          <CardActions
-            eventId={fallbackEventId}
-            profilePubkey={pubkey}
-            eventKind={0}
-            onToggleMenu={onToggleMenu}
-            menuButtonRef={menuButtonRef}
-          />
-        </div>
-      </div>
-    </div>
-  );
-}
+import ExplorerPortalMenu from '@/components/ExplorerPortalMenu';
+import ProfileBanner from '@/components/ProfileBanner';
+import ProfileCreatedAt from '@/components/ProfileCreatedAt';
 
 type ProfileCardProps = {
   event: NDKEvent;
@@ -185,7 +30,6 @@ export default function ProfileCard({ event, onAuthorClick, onHashtagClick, show
   const profile = (event.author?.profile as ProfileLike);
   const bannerUrl = profile?.banner || profile?.cover || profile?.header;
   const safeBannerUrl = isAbsoluteHttpUrl(bannerUrl) ? bannerUrl : undefined;
-  const [bannerExpanded, setBannerExpanded] = useState(false);
   const router = useRouter();
   const pathname = usePathname();
   const [showRaw, setShowRaw] = useState(false);
@@ -247,83 +91,8 @@ export default function ProfileCard({ event, onAuthorClick, onHashtagClick, show
     <div className={noteCardClasses}>
       {/* Centralized preseed: whenever a ProfileCard renders, ensure its event is prepared and seeded */}
       {(() => { try { if (event?.kind === 0 && event?.author?.pubkey) { setPrefetchedProfile(event.author.pubkey, prepareProfileEventForPrefetch(event)); } } catch {} return null; })()}
-      {showBanner && safeBannerUrl && (
-        <div
-          role="button"
-          tabIndex={0}
-          onClick={() => setBannerExpanded((prev) => !prev)}
-          onKeyDown={(e) => {
-            if (e.key === 'Enter' || e.key === ' ') {
-              e.preventDefault();
-              setBannerExpanded((prev) => !prev);
-            }
-          }}
-          className="block w-full focus:outline-none"
-          aria-expanded={bannerExpanded}
-          title={bannerExpanded ? 'Collapse banner' : 'Expand banner'}
-        >
-          <div
-            className="relative w-full border-b border-[#3d3d3d]"
-            style={{ height: bannerExpanded ? 240 : 32 }}
-          >
-            <div className="absolute inset-0 overflow-hidden">
-              <Image src={safeBannerUrl} alt="Banner" fill className="object-cover" unoptimized />
-            </div>
-            <div className="absolute top-1 left-1 z-50 flex gap-1">
-              <TitleBarButton
-                icon={faArrowLeft}
-                title="Go back"
-                onClick={() => router.back()}
-              />
-            </div>
-            <div className="absolute top-1 right-1 flex gap-1">
-              <TitleBarButton
-                title="Minimize"
-                textSize="text-[10px]"
-                onClick={() => setBannerExpanded(false)}
-              >
-                –
-              </TitleBarButton>
-              <TitleBarButton
-                title="Maximize"
-                textSize="text-[10px]"
-                onClick={() => {
-                  if (safeBannerUrl) window.open(safeBannerUrl, '_blank', 'noopener,noreferrer');
-                }}
-              >
-                ▢
-              </TitleBarButton>
-              <TitleBarButton
-                title="Close"
-                textSize="text-[10px]"
-                onClick={() => router.push('/')}
-              >
-                ×
-              </TitleBarButton>
-            </div>
-          </div>
-        </div>
-      )}
-      {showBanner && !bannerUrl && (
-        <div className="relative w-full border-b border-[#3d3d3d] bg-[#2d2d2d] rounded-t-lg" style={{ height: 32 }}>
-          <div className="absolute top-1 left-1 z-50 flex gap-1">
-            <TitleBarButton
-              icon={faArrowLeft}
-              title="Go back"
-              onClick={() => router.back()}
-            />
-          </div>
-          <div className="absolute top-1 right-1 flex gap-1">
-            <TitleBarButton
-              title="Close"
-              textSize="text-[10px]"
-              onClick={() => router.push('/')}
-            >
-              ×
-            </TitleBarButton>
-          </div>
-        </div>
-      )}
+      {showBanner && safeBannerUrl && <ProfileBanner bannerUrl={safeBannerUrl} />}
+      {showBanner && !bannerUrl && <ProfileBanner />}
       <div className="p-4">
       <div className="flex items-center justify-between">
         <div className="flex items-center gap-4">
@@ -434,78 +203,20 @@ export default function ProfileCard({ event, onAuthorClick, onHashtagClick, show
         }}
         menuButtonRef={portalButtonRef}
       />
-      
-      {showPortalMenu && typeof window !== 'undefined' && event?.author?.npub && createPortal(
-        <>
-          <div
-            className="fixed inset-0 z-[9998]"
-            onClick={(e) => { e.preventDefault(); setShowPortalMenu(false); }}
+
+      {showPortalMenu && event?.author?.npub && (() => {
+        const items = createProfileExplorerItems(event.author.npub, event.author.pubkey);
+        return (
+          <ExplorerPortalMenu
+            position={menuPosition}
+            onClose={() => setShowPortalMenu(false)}
+            portalItems={items.slice(0, -2)}
+            clientItems={items.slice(-2)}
+            showRaw={showRaw}
+            onToggleRaw={() => setShowRaw(v => !v)}
           />
-          <div
-            className="absolute z-[9999] w-56 rounded-md bg-[#2d2d2d]/95 border border-[#3d3d3d] shadow-lg backdrop-blur-sm"
-            style={{ top: menuPosition.top, left: menuPosition.left }}
-            onClick={(e) => { e.stopPropagation(); }}
-          >
-            <ul className="py-1 text-sm text-gray-200">
-              {(() => {
-                const items = createProfileExplorerItems(event.author.npub, event.author.pubkey);
-                const portalItems = items.slice(0, -2); // All items except last two
-                const clientItems = items.slice(-2); // Last two items (Web Client, Native App)
-                
-                return (
-                  <>
-                    {portalItems.map((item) => (
-                      <li key={item.name}>
-                        <a
-                          href={item.href}
-                          target={item.href.startsWith('http') ? '_blank' : undefined}
-                          rel={item.href.startsWith('http') ? 'noopener noreferrer' : undefined}
-                          className="px-3 py-2 hover:bg-[#3a3a3a] flex items-center justify-between"
-                          onClick={(e) => { e.stopPropagation(); setShowPortalMenu(false); }}
-                        >
-                          <span>{item.name}</span>
-                          <FontAwesomeIcon icon={faArrowUpRightFromSquare} className="text-gray-400 text-xs" />
-                        </a>
-                      </li>
-                    ))}
-                    <li className="border-t border-[#3d3d3d] my-1"></li>
-                    <li>
-                      <button
-                        type="button"
-                        className="w-full text-left px-3 py-2 hover:bg-[#3a3a3a] flex items-center justify-between"
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          setShowRaw(v => !v);
-                          setShowPortalMenu(false);
-                        }}
-                      >
-                        <span>{showRaw ? 'Hide raw JSON' : 'Show raw JSON'}</span>
-                        <FontAwesomeIcon icon={faCode} className="text-gray-400 text-xs" />
-                      </button>
-                    </li>
-                    <li className="border-t border-[#3d3d3d] my-1"></li>
-                    {clientItems.map((item) => (
-                      <li key={item.name}>
-                        <a
-                          href={item.href}
-                          target={item.href.startsWith('http') ? '_blank' : undefined}
-                          rel={item.href.startsWith('http') ? 'noopener noreferrer' : undefined}
-                          className="px-3 py-2 hover:bg-[#3a3a3a] flex items-center justify-between"
-                          onClick={(e) => { e.stopPropagation(); setShowPortalMenu(false); }}
-                        >
-                          <span>{item.name}</span>
-                          <FontAwesomeIcon icon={item.name === 'Native App' ? faMobileScreenButton : faArrowUpRightFromSquare} className="text-gray-400 text-xs" />
-                        </a>
-                      </li>
-                    ))}
-                  </>
-                );
-              })()}
-            </ul>
-          </div>
-        </>,
-        document.body
-      )}
+        );
+      })()}
     </div>
   );
 }
